src/test/java/com/phasin/invoices/LocalPaymentCreatorTest.java
package com.phasin.invoices;

import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.phasin.banking.Payment;
import com.phasin.banking.PaymentCreator;
import com.phasin.banking.PaymentException;
import com.phasin.invoices.Invoice;
import com.phasin.invoices.LocalPaymentCreator;

import java.math.BigDecimal;

import static org.testng.Assert.assertEquals;

public class LocalPaymentCreatorTest {

    PaymentCreator paymentCreator;

    @BeforeMethod
    void setUp() {
        paymentCreator = new LocalPaymentCreator();
    }

    @Test
    public void testCreatePayment() throws PaymentException {
        Invoice invoice = new Invoice(null, "some-address", "some-account", BigDecimal.TEN);
        Payment payment = paymentCreator.createPayment(invoice);
        assertEquals(payment.getReceiverAccount(), invoice.getAccount());
    }

    @Test(expectedExceptions = PaymentException.class)
    public void testCreatePayment_WithoutAccount() throws PaymentException {
        Invoice invoice = new Invoice("some-iban", "some-address", null, BigDecimal.TEN);
        paymentCreator.createPayment(invoice);
    }

}
